requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# GraphQL
strawberry-graphql[fastapi]==0.217.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
aiosqlite==0.19.0

# Utilities
python-dateutil==2.8.2
